move getDraftPicks
